+++ ClientDriver_OSVR.h
@@ -30,6 +30,8 @@
 
 // Library/third-party includes
 #include <openvr_driver.h>
+#include <osvr_compiler_detection.h>
+#include <string>
 
 // Standard includes
 // - none
@@ -99,52 +101,5 @@ private:
     std::string userDriverConfigDir_;
     std::string driverInstallDir_;
 };
-
-vr::EVRInitError ClientDriver_OSVR::Init(vr::IDriverLog* driver_log, vr::IClientDriverHost* driver_host, const char* user_driver_config_dir, const char* driver_install_dir)
-{
-    logger_ = driver_log;
-    driverHost_ = driver_host;
-    userDriverConfigDir_ = user_driver_config_dir;
-    driverInstallDir_ = driver_install_dir;
-
-    // TODO ?
-
-    return vr::VRInitError_None;
-}
-
-void ClientDriver_OSVR::Cleanup()
-{
-    logger_ = nullptr;
-    driverHost_ = nullptr;
-    userDriverConfigDir_.clear();
-    driverInstallDir_.clear();
-}
-
-bool ClientDriver_OSVR::BIsHmdPresent(const char* user_config_dir)
-{
-    // TODO
-    return true;
-}
-
-vr::EVRInitError ClientDriver_OSVR::SetDisplayId(const char* display_id)
-{
-    // TODO
-    return vr::VRInitError_None;
-}
-
-vr::HiddenAreaMesh_t ClientDriver_OSVR::GetHiddenAreaMesh(vr::EVREye eye)
-{
-    vr::HiddenAreaMesh_t hidden_area_mesh;
-    hidden_area_mesh.pVertexData = nullptr;
-    hidden_area_mesh.unTriangleCount = 0;
-
-    return hidden_area_mesh;
-}
-
-uint32_t ClientDriver_OSVR::GetMCImage(uint32_t *pImgWidth, uint32_t *pImgHeight, uint32_t *pChannels, void *pDataBuffer, uint32_t unBufferLen )
-{
-    // what is this function for? SteamVR drivers return 0, doing the same here.
-    return 0;
-}
 
 #endif // INCLUDED_ClientDriver_OSVR_h_GUID_7C0E8547_F8CF_4186_B637_9488CD6E3663
