+++ ServerDriver_OSVR.h
@@ -104,77 +104,4 @@ private:
     vr::IDriverLog* logger_;
 };
 
-vr::EVRInitError ServerDriver_OSVR::Init(vr::IDriverLog* driver_log, vr::IServerDriverHost* driver_host, const char* user_driver_config_dir, const char* driver_install_dir)
-{
-    logger_ = driver_log;
-
-    logger_->Log("ServerDriver_OSVR::Init() called.\n");
-    context_ = std::make_unique<osvr::clientkit::ClientContext>("com.osvr.SteamVR");
-
-    const std::string display_description = context_->getStringParameter("/display");
-    trackedDevices_.emplace_back(std::make_unique<OSVRTrackedDevice>(display_description, *(context_.get()), driver_host, logger_));
-
-    return vr::VRInitError_None;
-}
-
-void ServerDriver_OSVR::Cleanup()
-{
-    trackedDevices_.clear();
-    context_.reset();
-}
-
-uint32_t ServerDriver_OSVR::GetTrackedDeviceCount()
-{
-    std::string msg = "ServerDriver_OSVR::GetTrackedDeviceCount(): Detected " + std::to_string(trackedDevices_.size()) + " tracked devices.\n";
-    logger_->Log(msg.c_str());
-    return trackedDevices_.size();
-}
-
-vr::ITrackedDeviceServerDriver* ServerDriver_OSVR::GetTrackedDeviceDriver(uint32_t index)
-{
-    if (index >= trackedDevices_.size()) {
-        std::string msg = "ServerDriver_OSVR::GetTrackedDeviceDriver(): ERROR: Index " + std::to_string(index) + " is out of range [0.." + std::to_string(trackedDevices_.size()) + "].\n";
-        logger_->Log(msg.c_str());
-        return NULL;
-    }
-
-    std::string msg = "ServerDriver_OSVR::GetTrackedDeviceDriver(): Returning tracked device " + std::to_string(index) + ".\n";
-    logger_->Log(msg.c_str());
-    return trackedDevices_[index].get();
-}
-
-vr::ITrackedDeviceServerDriver* ServerDriver_OSVR::FindTrackedDeviceDriver(const char* id)
-{
-    for (auto& tracked_device : trackedDevices_) {
-        if (0 == std::strcmp(id, tracked_device->GetId())) {
-            std::string msg = "ServerDriver_OSVR::FindTrackedDeviceDriver(): Returning tracked device " + std::string(id) + ".\n";
-            logger_->Log(msg.c_str());
-            return tracked_device.get();
-        }
-    }
-
-    std::string msg = "ServerDriver_OSVR::FindTrackedDeviceDriver(): ERROR: Failed to locate device named '" + std::string(id) + "'.\n";
-    logger_->Log(msg.c_str());
-    return NULL;
-}
-
-void ServerDriver_OSVR::RunFrame()
-{
-    context_->update();
-}
-
-bool ServerDriver_OSVR::ShouldBlockStandbyMode()
-{
-    return false;
-}
-
-void ServerDriver_OSVR::EnterStandby()
-{
-}
-
-void ServerDriver_OSVR::LeaveStandby()
-{
-}
-
-
 #endif // INCLUDED_ServerDriver_OSVR_h_GUID_136B1359_C29D_4198_9CA0_1C223CC83B84
